refactor: png 파일 이동
