+++ src/screens/login/LoginScreen.tsx
@@ -1,7 +1,8 @@
-import LoginContainer from '@/components/login/loginContainer';
+import LoginContainer from '@/components/container/loginContainer';
 import KakaoLogin from '@/components/button/kakaoLogin';
 import {Image, Text, View} from 'react-native';
 import loginScreenStyles from '@/screens/login/LoginScreen.style';
+import {defaultImages} from '@/assets';
 
 function LoginScreen() {
   return (
@@ -11,7 +12,7 @@ function LoginScreen() {
           {`영잘알들의
 영화교류 플랫폼`}
         </Text>
-        <Image source={require('@/assets/popcornmate.png')} />
+        <Image source={defaultImages.popCornMate} />
       </View>
       <View style={loginScreenStyles.wrapper}>
         <KakaoLogin />
